(svn r17487) -Fix: silence some MSVC warnings

+++ src/company_cmd.cpp
@@ -77,8 +77,8 @@ Company::~Company()
  */
 void Company::PostDestructor(size_t index)
 {
-	InvalidateWindowData(WC_GRAPH_LEGEND, 0, index);
-	InvalidateWindowData(WC_PERFORMANCE_DETAIL, 0, index);
+	InvalidateWindowData(WC_GRAPH_LEGEND, 0, (int)index);
+	InvalidateWindowData(WC_PERFORMANCE_DETAIL, 0, (int)index);
 }
 
 /**
